+++ order_food_online_project/notes/views.py
@@ -21,15 +21,19 @@ def note_search(request):
 
 
 def note_add(request):
-    if request.method == "POST":
-        form = NoteAddForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect(reverse('note_list'))
+    if request.user.has_perm('notes.add_note'):
+        if request.method == "POST":
+            form = NoteAddForm(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect(reverse('note_list'))
 
-    form = NoteAddForm()
+        form = NoteAddForm()
 
-    return render(request, 'notes/note_add.html', context={'form': form})
+        return render(request, 'notes/note_add.html', context={'form': form})
+    else:
+        request.session['permission_codename'] = 'add_note'
+        return HttpResponseRedirect(reverse('permission_denied'))
 
 
 def note_view(request, note_id):
@@ -39,24 +43,32 @@ def note_view(request, note_id):
 
 
 def note_edit(request, note_id):
-    note = Note.objects.get(id=note_id)
-    if note:
+    if request.user.has_perm('notes.change_note'):
+        note = Note.objects.get(id=note_id)
+        if note:
 
-        if request.method == "POST":
-            form = NoteAddForm(request.POST, request.FILES, instance=note)
-            if form.is_valid():
-                form.save()
-                return HttpResponseRedirect(reverse('note_list'))
+            if request.method == "POST":
+                form = NoteAddForm(request.POST, request.FILES, instance=note)
+                if form.is_valid():
+                    form.save()
+                    return HttpResponseRedirect(reverse('note_list'))
 
-        form = NoteAddForm()
+            form = NoteAddForm()
 
-        return render(request, 'notes/note_edit.html', context={'note': note, 'form': form})
+            return render(request, 'notes/note_edit.html', context={'note': note, 'form': form})
+    else:
+        request.session['permission_codename'] = 'change_note'
+        return HttpResponseRedirect(reverse('permission_denied'))
 
 
 def note_delete(request, note_id):
-    note = Note.objects.get(id=note_id)
+    if request.user.has_perm('notes.delete_note'):
+        note = Note.objects.get(id=note_id)
 
-    if note:
-        if request.method == "POST":
-            note.delete()
-            return HttpResponseRedirect(reverse('note_list'))
+        if note:
+            if request.method == "POST":
+                note.delete()
+                return HttpResponseRedirect(reverse('note_list'))
+    else:
+        request.session['permission_codename'] = 'delete_note'
+        return HttpResponseRedirect(reverse('permission_denied'))
